move email & password validation to services
